+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 # CHANGELOG
 
+## 3 : mongodb id check
+- added a function for checking the validity of mongodb object IDs
+- removed some slight redundancy in the error message for checking a UUID
+
 ## 2 : string long fix
 - fixed an issue where `checkString()` was converting the `options.max` value to a boolean